implement directory traversal
- add more file manipulation helpers to main
(scanDirectory and joinPath). joinPath is needed
because path APIs are blocked in the browser.
scanDirectory returns a list of directories and mkv files.
- name navigation component reactive, relecting the current
directory data and adding onclick handlers to the directory
items to update the current directory.
- remove static test data for directory/file items.
static file detail data is still present as a placeholder
- consoliate type definitions for state data into data.svelte.ts,
eliminating data.ts. should data.svelte.ts be renamed to
state.svelte.ts?