remove unused tailwind attributes
when originally copying some of the layout from the
shadcn mail template, some extra attributes were carried over.
this includes some 'dark mode' and 'collapsed component' attributes.
neither of these are currently used. dark mode can be added back
later if there's a need, but it needs to be validated.