fix renderer intellisense
VSCode has been showing errors that shadcn component modules can't
be found (such as "$lib/components/ui/button").
to troubleshoot this, the path alias were moved to tsconfig.web.json,
where they arguably should have been put when I added shadcn support,
since shadcn is only used by the renderer (web) part of the project
anyway. It appears that the real fix was to remove the '$lib' path,
leaving only the '$lib/*' one. This was never a problem for the
build, only intellisense, but it was annoying.

I also added the 'composite=true' flag to the web config, since
VSCode was complaining about it. not sure whether it's truly needed, but
it doesn't seem to cause issues.