fix for intellisense for $lib
after a few attempts to fix this, with various side effects,
hopefully this is makes all of the tools properly recognize $lib.

After some troubleshooting, my observations are:
- electron.vite.config.ts aliases are used for the real build
when running the electron app
- tsconfig.json is used by the shadcn command line
tool (npx shadcn-svelte@latest add ...)
- tsconfig.web.json is used by vscode/typescript intellisense

to make all of the above work, the $lib alias needs to be in all 3
spots. sigh.