bug fixes
- fix display of raw details. previously, the textarea was only resized
to fit the content when the file selection was changed, so selecting
a file then selecting 'raw' would lead to a 0-height (therefore hidden)
textarea. this adds a resize call to the tab selection, too, so the
raw details always display correctly.
- add highlighting to the button for the currently selected file
in the file list
- change the tooltip which displays the full file/directory name for
navigation list items. it now appears when hovering anywhere over the
button, not just over the text.