fix display of raw details
textarea was originally used because it seemed like the most
staightforward way of laying out pre-formatted text.
unfortunately, textarea doesn't size to its content, and doesn't
expose a lot of the events or attributes that most elements expose.
This makes it hard to use in a reactive layout.

Turns out, using a div with writespace-pre-wrap and a mono-space font
provides the formatting textarea does, and has all of the reactive
behavior that textarea doesn't.

YAY!