Fix overflow of page width
After a lot of experimentation, the width overflow was at the app level.
It was because the flexbox at the top level of the body wasn't limiting
the resize group to the page, even with flex-shrink attributes and width
attributes on everything. At best, I could get flexbox to only overflow
by the width of the button strip, even with the button strip inside the
flexbox. Frustrating. Switching to grid made everything work properly.
I still think it should be possible to make flexbox work, but I'm not
sure how.

Additionally, get rid of the 'defaultLayout' array. the defaultSize
attributes are relative fractions, and using 1:2 seems entirely
reasonable for the layout.