add resizable pane at bottom of app for displaying messages
this is a checkpoint. after many days of research and troubleshooting,
this change finally gets the pane added with the right set of divs and
width attributes to get scrollbars working properly across all of the
different panes.

flexbox doesn't seem to like items with scrollable content in the middle,
but grid seems to handle it better. it also seems that this whole parent
chain of the scrollable item needs h-full classes so that each item takes
the entire height of its parent. And sometimes items need an extra
h-full div added as a container. it's still a little mystifying
how the items interact to lay things out.